Add per-page metadata and remove head tags
