--- app/impressum/head.tsx
@@ -1,18 +1,0 @@
-export default function Head() {
-  return (
-    <>
-      <title>Impressum & Datenschutz | Mathis Kräkel</title>
-      <meta
-        name="description"
-        content="Impressum und Datenschutzerklärung von Mathis Kräkel."
-      />
-      <link rel="canonical" href="https://mathiskraekel.de/impressum" />
-      <meta property="og:title" content="Impressum & Datenschutz | Mathis Kräkel" />
-      <meta property="og:description" content="Impressum und Datenschutzerklärung von Mathis Kräkel." />
-      <meta property="og:type" content="website" />
-      <meta property="og:url" content="https://mathiskraekel.de/impressum" />
-      <meta property="og:site_name" content="Mathis Kräkel – Softwareentwicklung & Beratung" />
-      <meta name="robots" content="index, follow" />
-    </>
-  );
-}
