--- app/lead-generation/head.tsx
@@ -1,18 +1,0 @@
-export default function Head() {
-  return (
-    <>
-      <title>Intelligente Lead-Formulare | Mathis Kräkel</title>
-      <meta
-        name="description"
-        content="Maßgeschneiderte Formulare mit Echtzeit-Berechnungen und automatischer Lead-Qualifizierung."
-      />
-      <link rel="canonical" href="https://mathiskraekel.de/lead-generation" />
-      <meta property="og:title" content="Intelligente Lead-Formulare | Mathis Kräkel" />
-      <meta property="og:description" content="Maßgeschneiderte Formulare mit Echtzeit-Berechnungen und automatischer Lead-Qualifizierung." />
-      <meta property="og:type" content="website" />
-      <meta property="og:url" content="https://mathiskraekel.de/lead-generation" />
-      <meta property="og:site_name" content="Mathis Kräkel – Softwareentwicklung & Beratung" />
-      <meta name="robots" content="index, follow" />
-    </>
-  );
-}
