--- app/mvp-development/head.tsx
@@ -1,18 +1,0 @@
-export default function Head() {
-  return (
-    <>
-      <title>MVP Development in Rekordzeit | Mathis Kräkel</title>
-      <meta
-        name="description"
-        content="Schnelle Entwicklung Ihres Minimum Viable Products mit modernem Tech-Stack. Von der Idee bis zum Launch in 4-8 Wochen."
-      />
-      <link rel="canonical" href="https://mathiskraekel.de/mvp-development" />
-      <meta property="og:title" content="MVP Development in Rekordzeit | Mathis Kräkel" />
-      <meta property="og:description" content="Schnelle Entwicklung Ihres Minimum Viable Products mit modernem Tech-Stack. Von der Idee bis zum Launch in 4-8 Wochen." />
-      <meta property="og:type" content="website" />
-      <meta property="og:url" content="https://mathiskraekel.de/mvp-development" />
-      <meta property="og:site_name" content="Mathis Kräkel – Softwareentwicklung & Beratung" />
-      <meta name="robots" content="index, follow" />
-    </>
-  );
-}
